web/news/models.py
from datetime import datetime

from flask_sqlalchemy import SQLAlchemy
import uuid64


db = SQLAlchemy()


class CRUDMixin(object):
    """Copied from https://realpython.com/blog/python/python-web-applications-with-flask-part-ii/
    """  # noqa

    __table_args__ = {'extend_existing': True}

    id = db.Column(db.BigInteger, primary_key=True, autoincrement=False,
                   default=uuid64.issue())

    @classmethod
    def create(cls, commit=True, **kwargs):
        if 'id' not in kwargs:
            kwargs.update(dict(id=uuid64.issue()))
        instance = cls(**kwargs)

        if hasattr(instance, 'timestamp') \
                and getattr(instance, 'timestamp') is None:
            instance.timestamp = datetime.utcnow()

        return instance.save(commit=commit)

    @classmethod
    def get(cls, id):
        return cls.query.get(id)

    # We will also proxy Flask-SqlAlchemy's get_or_404 for symmetry
    @classmethod
    def get_or_404(cls, id):
        return cls.query.get_or_404(id)

    @classmethod
    def exists(cls, **kwargs):
        row = cls.query.filter_by(**kwargs).first()
        return row is not None

    def update(self, commit=True, **kwargs):
        for attr, value in kwargs.iteritems():
            setattr(self, attr, value)
        return commit and self.save() or self

    def save(self, commit=True):
        db.session.add(self)
        if commit:
            db.session.commit()
        return self

    def delete(self, commit=True):
        db.session.delete(self)
        return commit and db.session.commit()


class Article(db.Model, CRUDMixin):
    published_at = db.Column(db.DateTime(timezone=False))
    fetched_at = db.Column(db.DateTime(timezone=False))

    #: Name of media (e.g., CNN, FoxNews, etc)
    channel = db.Column(db.String)

    #: Canonical URL
    url = db.Column(db.String)

    title = db.Column(db.String)
    content = db.Column(db.Text)
    summary = db.Column(db.Text)


# TODO: Consider keyword-article mappings
